+++ docs/changelog.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+* `cols_move` method for moving table columns after another column.
 * Pipe method for applying functions to `TypTable` objects in a method chain.
 
 ## 0.1.4 - 2026-05-10